Use maven for dependency management
