+++ src/main/java/modreq/MessageType.java
