+++ src/main/java/modreq/korik/Configurable.java
