+++ src/main/java/modreq/managers/CommandManager.java
